LMN/requirements.txt
Django
Pillow
psycopg2-binary
djangorestframework